Fix imports, typos, move tests
