+++ rasa/core/channels/__init__.py
@@ -25,6 +25,7 @@ from rasa.core.channels.slack import SlackInput  # nopep8
 from rasa.core.channels.telegram import TelegramInput  # nopep8
 from rasa.core.channels.twilio import TwilioInput  # nopep8
 from rasa.core.channels.webexteams import WebexTeamsInput  # nopep8
+from rasa.core.channels.hangouts import HangoutsInput  # nopep8
 
 input_channel_classes = [
     CmdlineInput,
@@ -40,6 +41,7 @@ input_channel_classes = [
     RestInput,
     SocketIOInput,
     WebexTeamsInput,
+    HangoutsInput,
 ]  # type: List[InputChannel]
 
 # Mapping from a input channel name to its class to allow name based lookup.
